+++ CHANGELOG.md
@@ -10,10 +10,12 @@ and this library adheres to Rust's notion of
 - `zcash_spec::PrfExpand::REGISTERED_ZIP32_CHILD` (for tagged ZIP 32 child
   derivation).
 - `zcash_spec::PrfExpand::ADHOC_ZIP32_CHILD` (renamed and retyped per ZIP 32).
+- `zcash_spec::VariableLengthSlice`
 
 ### Changed
 - `zcash_spec::PrfExpand::ORCHARD_ZIP32_CHILD` now has type
-  `PrfExpand<([u8; 32], [u8; 4], Option<(u8, &[u8])>)>` due to ZIP 32 changes.
+  `PrfExpand<([u8; 32], [u8; 4], Option<(u8, VariableLengthSlice)>)>` due to
+  ZIP 32 changes.
 
 ### Removed
 - `zcash_spec::PrfExpand::ARBITRARY_ZIP32_CHILD` (use `ADHOC_ZIP32_CHILD`